corrected long-standing blunders in MCFSimplex (but not all of them)
opening and closing arcs had serious issues; furthermore, the code had a bunch of ugly things that screamed for mending. I did something but some work would still be needed for which I do not have time. In particular, reoptimization was completely disabled: now it is enabled but off by default as it has issues that would require fixing.
I'll open an issue to keep track of it, but I have no time for this now.